backend/pyproject.toml
[project]
name = "nextjs-fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.120.4",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
]
